# Use a slotmap for render assets storage