… crates/bevy_pbr/src/meshlet/gpu_scene.rs
… crates/bevy_pbr/src/meshlet/material_draw_prepare.rs
… crates/bevy_pbr/src/prepass/mod.rs
… crates/bevy_pbr/src/render/light.rs
… crates/bevy_pbr/src/render/mesh.rs
… crates/bevy_render/src/extract_instances.rs
… crates/bevy_sprite/src/mesh2d/material.rs
… crates/bevy_sprite/src/mesh2d/mesh.rs
… crates/bevy_utils/Cargo.toml
… examples/2d/mesh2d_manual.rs
… examples/shader/shader_instancing.rs