# Objective

- Optimise material mesh entity rendering code paths
- Prepare for 'draw stream' rendering which needs to pack render commands into a `Vec<u32>`

## Solution

- AssetId is quite large due to supporting a UUID variant and the type is used in very hot loops for batching and rendering. In rendering, we don't need the UUID. We can use a smaller id to refer to the prepared render asset.
- Use a slotmap as the backing store for `RenderAssets`
- Maintain a map from `AssetId<A: RenderAsset>` to slotmap key inside `RenderAssets` and use it to look up keys for known render assets at extraction time
- Add a `UpdatePendingRenderAssetKeyPlugin` to help with updating asset keys on instance data for assets that were loaded during the current frame, to avoid a 1-frame lag.

## Benchmarks

On an M1 Max, running `cargo run --release --example bevymark -- --benchmark --waves 160 --per-wave 1000 --mode mesh2d`. Yellow is main, red is this PR:

<img width="804" alt="Screenshot 2024-04-18 at 12 20 21" src="https://github.com/bevyengine/bevy/assets/302146/893c24aa-29a6-48dc-9107-ff76f8a6d96e">

A 6.3% reduction in median frame time.

Main:
<img width="1165" alt="Screenshot 2024-04-18 at 12 20 54" src="https://github.com/bevyengine/bevy/assets/302146/4aaec06d-d660-44ec-b972-7dafefe0994c">

PR:
<img width="1160" alt="Screenshot 2024-04-18 at 12 21 19" src="https://github.com/bevyengine/bevy/assets/302146/10eec853-ee77-4719-aa07-647a8994e9bc">

Noteworthy spans:
```
92.16ms -> 86.00ms : main_pass_2d
19.20ms -> 17.72mz : batch_and_prepare_sorted_render_phase
 6.07ms ->  4.34ms : queue_material2d_meshes
 0.95ms ->  1.71ms : extract_material_meshes_2d
```

---

## Changelog

- Added: `RenderAssetKey` slotmap key type to be used with `RenderAssets`
- Added: `get_key`, `get_with_key`, and `get_with_key_mut` methods to `RenderAssets` to use slotmap keys.
- Added: `UpdatePendingRenderAssetKeyPlugin` to help with updating asset keys on instance data for assets that were loaded during the current frame, to avoid a 1-frame lag.
- Changed: Use slotmap for internal storage for `RenderAssets` for improved performance due to the smaller asset reference data size of slotmap keys compared to `AssetId`