@pcwalton 2024-04-20
@cart Weren't you planning on removing `AssetId::Uuid` at some point?