@superdump 2024-04-21
I'm moving this PR to a draft. After the push back about the regression in extraction time, I went down a rabbit hole of changing Handle and AssetId to not be enums, instead to be structs and require always having an AssetIndex member, and optionally a Uuid. See Discord for details, but the short version is it performs even better than this PR, and it feels like a better solution if we can stomach not having support for compile-time const Handle.