@superdump 2024-04-27
Closing. After discussion with @cart we decided not to do this, but rather wait for the long-term solution.